+++ models/providers/ConfigurableReferenceProvider.js
@@ -1,38 +1,40 @@
 const vscode = require('vscode');
-const ConfigurableDefinitionOptions = require('../options/ConfigurableDefinitionOptions');
+const ConfigurableReferenceOptions = require('../options/ConfigurableReferenceOptions');
 const PathProvider = require('../services/PathProvider');
 const Logger = require('../services/Logger');
 const fs = require('fs').promises;
 
-module.exports = class ConfigurableDefinitionProvider {
+module.exports = class ConfigurableReferenceProvider {
 
 	/**
-	 * @type {ConfigurableDefinitionOptions}
+	 * @type {ConfigurableReferenceOptions}
 	 */
 	options;
 
 	/**
-	 * @param {ConfigurableDefinitionOptions} options 
+	 * @param {ConfigurableReferenceOptions} options 
 	 */
 	constructor(options) {
 		this.options = options;
 	}
+
 	/**
 	 * @param {vscode.TextDocument} document 
 	 * @param {vscode.Position} position 
+	 * @param {vscode.ReferenceContext} context
 	 * @param {vscode.CancellationToken} token 
-	 * @returns {Promise<vscode.Location|undefined>}
+	 * @returns {Promise<Array<vscode.Location>|undefined>}
 	 */
-	async provideDefinition(document, position, token) {
+	async provideReferences(document, position, context, token) {
 
-		const definitionName = this.getDefinitionName(document, position);
-		if (!definitionName) {
+		const referenceName = this.getReferenceName(document, position);
+		if (!referenceName) {
 			//Couldn't extract a definition name from current location, just return
 			return;
 		}
 
-		const definitionResult = await this.findDefinitionInFiles(definitionName, document, token);
-		return definitionResult;
+		const referencesResult = await this.findReferencesInFiles(referenceName, document, token);
+		return referencesResult;
 	}
 
 	/**
@@ -40,10 +42,10 @@ module.exports = class ConfigurableDefinitionProvider {
 	 * @param {vscode.Position} position
 	 * @returns {string|undefined}
 	 */
-	getDefinitionName(document, position) {
+	getReferenceName(document, position) {
 		const currentLine = document.lineAt(position.line);
 		let match;
-		const regexp = new RegExp(this.options.definitionRegexp.source, this.options.definitionRegexp.flags);
+		const regexp = new RegExp(this.options.referenceRegexp.source, this.options.referenceRegexp.flags);
 		while (match = regexp.exec(currentLine.text)) {
 			if (position.character >= match.index && position.character <= match.index + match[0].length) {
 				const definitionName = match[1] || match[0];
@@ -58,31 +60,40 @@ module.exports = class ConfigurableDefinitionProvider {
 	 * @param {string} definitionName 
 	 * @param {vscode.TextDocument} document
 	 * @param {vscode.CancellationToken} token 
-	 * @returns {Promise<vscode.Location|undefined>}
+	 * @returns {Promise<Array<vscode.Location>|undefined>}
 	 */
-	async findDefinitionInFiles(definitionName, document, token) {
+	async findReferencesInFiles(definitionName, document, token) {
 
 		const includeGlobPattern = PathProvider.replaceDocumentPathVariables(this.options.includeGlobPattern, document);
 		const excludeGlobPattern = PathProvider.replaceDocumentPathVariables(this.options.excludeGlobPattern, document);
 		if (!includeGlobPattern) {
 			return;
 		}
-		const results = await vscode.workspace.findFiles(includeGlobPattern, excludeGlobPattern, undefined, token);
+		const results = await vscode.workspace.findFiles(includeGlobPattern, excludeGlobPattern, this.options.maxFiles, token);
 
 		if (results.length == 0) {
-			Logger.warn(`Couldn't find any file for include pattern ${includeGlobPattern} and exclude pattern ${excludeGlobPattern} in definition provider rule with definition regexp ${this.options.definitionRegexp}`);
+			Logger.warn(`Couldn't find any file for include pattern ${includeGlobPattern} and exclude pattern ${excludeGlobPattern} in reference provider rule with reference regexp ${this.options.referenceRegexp}`);
 			return;
 		}
+
+		/**
+		 * @type {Array<vscode.Location>}
+		 */
+		const locations = [];
 
 		for (const result of results) {
 			if (token.isCancellationRequested) {
 				return;
 			}
+			if (locations.length >= this.options.maxReferences) {
+				break;
+			}
 			const content = await fs.readFile(result.fsPath);
 			const textContent = content.toString();
 			const regexp = new RegExp(this.options.contentRegexp.source, this.options.contentRegexp.flags);
 			let match;
-			while (match = regexp.exec(textContent)) {
+			let referencesInFile = 0;
+			while ((match = regexp.exec(textContent)) && referencesInFile <= this.options.maxReferencesPerFile) {
 				const resultName = match[1] || match[0];
 				if (resultName != definitionName) {
 					continue;
@@ -93,8 +104,11 @@ module.exports = class ConfigurableDefinitionProvider {
 				const uri = vscode.Uri.file(result.fsPath);
 				const position = new vscode.Position(lines.length-1, lines[lines.length-1].length);
 				const location = new vscode.Location(uri, position);
-				return location;
+				locations.push(location);
+				referencesInFile++;
 			}
 		}
+
+		return locations;
 	}
 };