+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 
 ## [Unreleased]
 
-## [1.1.0] - 2017-07-28
+## [1.1.0] - 2017-07-29
 ### Added
+ - Reference providers.
  - Support for `${definition...}` placeholders in glob patterns for completion item providers.
- - Support hot configuration reload.
+ - Hot configuration reload.
 
 ## [1.0.0] - 2017-07-26
 ### Added